+++ cmd/cortextool/pkg/analyse/grafana.go
@@ -1,6 +1,7 @@
 package analyse
 
 import (
+	"encoding/json"
 	"fmt"
 	"regexp"
 	"sort"
@@ -114,14 +115,49 @@ func metricsFromTemplating(templating sdk.Templating, metrics map[string]struct{
 	return parseErrors
 }
 
+// Workaround to support Grafana "timeseries" panel. This should
+// be implemented in grafana/tools-sdk, and removed from here.
+func getCustomPanelTargets(panel sdk.Panel) *[]sdk.Target {
+	if panel.CommonPanel.Type != "timeseries" {
+		return nil
+	}
+
+	// Heavy handed approach to re-marshal the panel and parse it again
+	// so that we can extract the 'targets' field in the right format.
+
+	bytes, err := json.Marshal(panel.CustomPanel)
+	if err != nil {
+		log.Debugln("msg", "panel re-marshalling error", "err", err)
+		return nil
+	}
+
+	type panelType struct {
+		Targets []sdk.Target `json:"targets,omitempty"`
+	}
+
+	var parsedPanel panelType
+	err = json.Unmarshal(bytes, &parsedPanel)
+	if err != nil {
+		log.Debugln("msg", "panel parsing error", "err", err)
+		return nil
+	}
+
+	return &parsedPanel.Targets
+}
+
 func metricsFromPanel(panel sdk.Panel, metrics map[string]struct{}) []error {
 	var parseErrors []error
 
-	if panel.GetTargets() == nil {
-		return parseErrors
+	targets := panel.GetTargets()
+	if targets == nil {
+		targets = getCustomPanelTargets(panel)
+		if targets == nil {
+			parseErrors = append(parseErrors, fmt.Errorf("unsupported panel type: %q", panel.CommonPanel.Type))
+			return parseErrors
+		}
 	}
 
-	for _, target := range *panel.GetTargets() {
+	for _, target := range *targets {
 		// Prometheus has this set.
 		if target.Expr == "" {
 			continue