+++ cmd/cortextool/pkg/commands/analyse_grafana_test.go
@@ -38,3 +38,18 @@ func TestParseMetricsInBoard(t *testing.T) {
 	analyse.ParseMetricsInBoard(output, board)
 	assert.Equal(t, dashboardMetrics, output.Dashboards[0].Metrics)
 }
+
+func TestParseMetricsInBoardWithTimeseriesPanel(t *testing.T) {
+	var board sdk.Board
+	output := &analyse.MetricsInGrafana{}
+	output.OverallMetrics = make(map[string]struct{})
+
+	buf, err := loadFile("testdata/timeseries.json")
+	require.NoError(t, err)
+
+	err = json.Unmarshal(buf, &board)
+	require.NoError(t, err)
+
+	analyse.ParseMetricsInBoard(output, board)
+	assert.Equal(t, []string{"my_lovely_metric"}, output.Dashboards[0].Metrics)
+}